requirements.txt
aiohttp>=3.9.0
certifi>=2023.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
